requirements.txt
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.12.0
scanpy>=1.9.0
anndata>=0.8.0
POT>=0.8.0
matplotlib>=3.5.0
seaborn>=0.11.0
PyYAML>=6.0
tqdm>=4.62.0
